\name{Npop}
\alias{Npop}
\title{
Data for forest population number of within different age class.
}
\description{
Forest population number of within different age class.
}
\usage{data("Npop")}
\format{
  A data frame with 11 observations on the following 3 variables from the forest population survey data
  \describe{
    \item{\code{rank}}{Rank of age class of forest population.}
    \item{\code{ageclass}}{Age class of forest population.}
    \item{\code{ax}}{Forest population number of within different age class.}
  }
}
\details{
Population number of within different age class.
}
\author{
Zongzheng Chai, chaizz@126.com
}
\examples{
data(Npop)
Npop
}
